Reduce padding/spacing for modal content and buttons, based on size mode
Also consolidate the modal <ButtonBar> into modal.tsx (since it's only
used there) to make sharing consistent spacing values easier.
