--- libs/ui/src/button_bar.test.tsx
@@ -1,8 +1,0 @@
-import { render } from '../test/react_testing_library';
-
-import { ButtonBar } from './button_bar';
-
-test('renders ButtonBar', () => {
-  const { container } = render(<ButtonBar>foo</ButtonBar>);
-  expect(container.firstChild).toMatchSnapshot();
-});
