--- libs/ui/src/button_bar.tsx
@@ -1,27 +1,0 @@
-import styled from 'styled-components';
-
-export const ButtonBar = styled('div')`
-  display: flex;
-  flex-wrap: wrap-reverse;
-  align-items: center;
-  justify-content: flex-end;
-  border-top: ${(p) => p.theme.sizes.bordersRem.hairline}rem solid
-    ${(p) => p.theme.colors.outline};
-  padding: 0.5rem max(${(p) => p.theme.sizes.minTouchAreaSeparationPx}px, 1rem);
-  gap: max(${(p) => p.theme.sizes.minTouchAreaSeparationPx}px, 0.5rem);
-
-  & > * {
-    flex-grow: 1;
-  }
-
-  & > *:first-child {
-    order: 2;
-    min-width: 40%;
-  }
-
-  & > *:only-child {
-    @media (min-width: 480px) {
-      flex-grow: initial;
-    }
-  }
-`;
