+++ libs/ui/src/index.ts
@@ -5,7 +5,6 @@ export * from './big_metric';
 export * from './bmd_paper_ballot';
 export * from './breadcrumbs';
 export * from './button';
-export * from './button_bar';
 export * from './button_list';
 export * from './card';
 export * from './cast_vote_records';
